// src/models/schemas/Bookmark.schema.ts
import { ObjectId } from 'mongodb'

type BookmarkConstructorType = {
  _id?: ObjectId
  user_id: ObjectId
  tweet_id: ObjectId
  created_at?: Date
}

export default class Bookmark {
  _id?: ObjectId
  user_id: ObjectId
  tweet_id: ObjectId
  created_at: Date

  constructor({ _id, user_id, tweet_id, created_at }: BookmarkConstructorType) {
    this._id = _id || new ObjectId()
    this.user_id = user_id
    this.tweet_id = tweet_id
    this.created_at = created_at || new Date()
  }
}